Do not show decimals

+++ cnchi/download/download_requests.py
@@ -176,9 +176,9 @@ class Download(object):
                                 bps = (completed_length // (time.clock() - start)) / 1024
                                 if bps > 1024:
                                     Mbps = bps / 1024
-                                    progress_text = "{0}% {1:.2f} Mbps".format(percent * 100, Mbps)
+                                    progress_text = "{0}% {1:.2f} Mbps".format(int(percent * 100), Mbps)
                                 else:
-                                    progress_text = "{0}% {1:.2f} bps".format(percent * 100, bps)
+                                    progress_text = "{0}% {1:.2f} bps".format(int(percent * 100), bps)
                                 self.queue_event('progress_bar_show_text', progress_text)
 
                             md5 = md5_hash.hexdigest()
